Moved implementation to 'Internal' folder and namespace, to make them pubternal
