+++ perf/gfoidl.Base64.Benchmarks/DecodeStringBenchmark.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using BenchmarkDotNet.Attributes;
+using gfoidl.Base64.Internal;
 
 namespace gfoidl.Base64.Benchmarks
 {
