+++ perf/gfoidl.Base64.Benchmarks/DecodeUtf8Benchmark.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Buffers;
 using BenchmarkDotNet.Attributes;
+using gfoidl.Base64.Internal;
 
 namespace gfoidl.Base64.Benchmarks
 {
