+++ source/gfoidl.Base64/Internal/Avx2Helper.cs
@@ -4,7 +4,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.Intrinsics;
 using System.Runtime.Intrinsics.X86;
 
-namespace gfoidl.Base64
+namespace gfoidl.Base64.Internal
 {
     internal static class Avx2Helper
     {
