+++ source/gfoidl.Base64/Internal/Base64Encoder/Base64Encoder.Decode.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Buffers;
-using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -12,21 +11,47 @@ using System.Runtime.Intrinsics.X86;
 // Scalar based on https://github.com/dotnet/corefx/tree/master/src/System.Memory/src/System/Buffers/Text
 // SSE2 based on https://github.com/aklomp/base64/tree/master/lib/arch/ssse3
 // AVX2 based on https://github.com/aklomp/base64/tree/master/lib/arch/avx2
-// Lookup and validation for SSE2 and AVX2 based on http://0x80.pl/notesen/2016-01-17-sse-base64-decoding.html#vector-lookup-pshufb
 
-namespace gfoidl.Base64
+namespace gfoidl.Base64.Internal
 {
-    partial class Base64UrlEncoder
+    partial class Base64Encoder
     {
-        public override int GetDecodedLength(ReadOnlySpan<byte> encoded) => this.GetDecodedLength(encoded.Length);
-        public override int GetDecodedLength(ReadOnlySpan<char> encoded) => this.GetDecodedLength(encoded.Length);
+        public override int GetDecodedLength(ReadOnlySpan<byte> encoded)
+        {
+            int maxLen = this.GetDecodedLength(encoded.Length);
+
+            ref byte tmp = ref MemoryMarshal.GetReference(encoded);
+            ref byte end = ref Unsafe.Add(ref tmp, encoded.Length - 1);
+
+            int padding = 0;
+
+            if (end == EncodingPad) padding++;
+            if (Unsafe.Subtract(ref end, 1) == EncodingPad) padding++;
+
+            return maxLen - padding;
+        }
+        //---------------------------------------------------------------------
+        public override int GetDecodedLength(ReadOnlySpan<char> encoded)
+        {
+            int maxLen = this.GetDecodedLength(encoded.Length);
+
+            ref char tmp = ref MemoryMarshal.GetReference(encoded);
+            ref char end = ref Unsafe.Add(ref tmp, encoded.Length - 1);
+
+            int padding = 0;
+
+            if (end == EncodingPad) padding++;
+            if (Unsafe.Subtract(ref end, 1) == EncodingPad) padding++;
+
+            return maxLen - padding;
+        }
         //---------------------------------------------------------------------
         internal int GetDecodedLength(int encodedLength)
         {
-            if ((uint)encodedLength >= int.MaxValue)
+            if (encodedLength < 0)
                 ThrowHelper.ThrowArgumentOutOfRangeException(ExceptionArgument.length);
 
-            return GetDataLen(encodedLength, out int _);
+            return (encodedLength >> 2) * 3;
         }
         //---------------------------------------------------------------------
         // PERF: can't be in base class due to inlining (generic virtual)
@@ -69,9 +94,7 @@ namespace gfoidl.Base64
         {
             uint sourceIndex = 0;
             uint destIndex   = 0;
-
-            int dataLength = GetDataLen(inputLength, out int base64Len, isFinalBlock);
-            int srcLength  = base64Len & ~0x3;      // only decode input up to the closest multiple of 4.
+            int srcLength    = inputLength & ~0x3;      // only decode input up to the closest multiple of 4.
 
             ref byte destBytes  = ref MemoryMarshal.GetReference(data);
 
@@ -116,7 +139,7 @@ namespace gfoidl.Base64
             int maxSrcLength = 0;
             int destLength   = data.Length;
 
-            if (destLength >= dataLength)
+            if (destLength >= this.GetDecodedLength(srcLength))
             {
                 maxSrcLength = srcLength - skipLastChunk;
             }
@@ -160,50 +183,89 @@ namespace gfoidl.Base64
             // if isFinalBlock is false, we will never reach this point
 
             // Handle last four bytes. There are 0, 1, 2 padding chars.
-            int numPaddingChars = base64Len - inputLength;
-            ref T lastFourStart = ref Unsafe.Add(ref src, srcLength - 4);
+            uint t0, t1, t2, t3;
 
-            if (numPaddingChars == 0)
+            if (typeof(T) == typeof(byte))
             {
-                int result = DecodeFour(ref lastFourStart, ref decodingMap);
-
-                if (result < 0) goto InvalidExit;
-                if (destIndex > destLength - 3) goto DestinationSmallExit;
-
-                WriteThreeLowOrderBytes(ref destBytes, destIndex, result);
-                sourceIndex += 4;
-                destIndex   += 3;
+                ref byte tmp = ref Unsafe.As<T, byte>(ref src);
+                t0 = Unsafe.Add(ref tmp, (IntPtr)(srcLength - 4));
+                t1 = Unsafe.Add(ref tmp, (IntPtr)(srcLength - 3));
+                t2 = Unsafe.Add(ref tmp, (IntPtr)(srcLength - 2));
+                t3 = Unsafe.Add(ref tmp, (IntPtr)(srcLength - 1));
             }
-            else if (numPaddingChars == 1)
+            else if (typeof(T) == typeof(char))
             {
-                int result = DecodeThree(ref lastFourStart, ref decodingMap);
+                ref char tmp = ref Unsafe.As<T, char>(ref src);
+                t0 = Unsafe.Add(ref tmp, (IntPtr)(srcLength - 4));
+                t1 = Unsafe.Add(ref tmp, (IntPtr)(srcLength - 3));
+                t2 = Unsafe.Add(ref tmp, (IntPtr)(srcLength - 2));
+                t3 = Unsafe.Add(ref tmp, (IntPtr)(srcLength - 1));
+            }
+            else
+            {
+                throw new NotSupportedException();  // just in case new types are introduced in the future
+            }
 
-                if (result < 0)
+            int i0 = Unsafe.Add(ref decodingMap, (IntPtr)t0);
+            int i1 = Unsafe.Add(ref decodingMap, (IntPtr)t1);
+
+            i0 <<= 18;
+            i1 <<= 12;
+
+            i0 |= i1;
+
+            if (t3 != EncodingPad)
+            {
+                int i2 = Unsafe.Add(ref decodingMap, (IntPtr)t2);
+                int i3 = Unsafe.Add(ref decodingMap, (IntPtr)t3);
+
+                i2 <<= 6;
+
+                i0 |= i3;
+                i0 |= i2;
+
+                if (i0 < 0)
+                    goto InvalidExit;
+
+                if (destIndex > destLength - 3)
+                    goto DestinationSmallExit;
+
+                WriteThreeLowOrderBytes(ref destBytes, destIndex, i0);
+                destIndex += 3;
+            }
+            else if (t2 != EncodingPad)
+            {
+                int i2 = Unsafe.Add(ref decodingMap, (IntPtr)t2);
+
+                i2 <<= 6;
+
+                i0 |= i2;
+
+                if (i0 < 0)
                     goto InvalidExit;
 
                 if (destIndex > destLength - 2)
                     goto DestinationSmallExit;
 
-                WriteTwoLowOrderBytes(ref destBytes, destIndex, result);
-                sourceIndex += 3;
-                destIndex   += 2;
+                Unsafe.Add(ref destBytes, (IntPtr)destIndex)       = (byte)(i0 >> 16);
+                Unsafe.Add(ref destBytes, (IntPtr)(destIndex + 1)) = (byte)(i0 >> 8);
+                destIndex += 2;
             }
             else
             {
-                int result = DecodeTwo(ref lastFourStart, ref decodingMap);
-
-                if (result < 0)
+                if (i0 < 0)
                     goto InvalidExit;
 
                 if (destIndex > destLength - 1)
                     goto DestinationSmallExit;
 
-                WriteOneLowOrderByte(ref destBytes, destIndex, result);
-                sourceIndex += 2;
-                destIndex   += 1;
+                Unsafe.Add(ref destBytes, (IntPtr)destIndex) = (byte)(i0 >> 16);
+                destIndex += 1;
             }
 
-            if (srcLength != base64Len)
+            sourceIndex += 4;
+
+            if (srcLength != inputLength)
                 goto InvalidExit;
 #if NETCOREAPP
         DoneExit:
@@ -222,7 +284,7 @@ namespace gfoidl.Base64
 
         NeedMoreDataExit:
             consumed = (int)sourceIndex;
-            written  = (int)destIndex;
+            written = (int)destIndex;
             return OperationStatus.NeedMoreData;
 
         InvalidExit:
@@ -245,12 +307,10 @@ namespace gfoidl.Base64
             ref T simdSrcEnd   = ref Unsafe.Add(ref src, (IntPtr)((uint)sourceLength - 45 + 1));
 
             // The JIT won't hoist these "constants", so help him
-            Vector256<sbyte> allOnes          = Avx.SetAllVector256<sbyte>(-1);     // -1 = 0xFF = true in simd
             Vector256<sbyte> lutHi            = s_avx_decodeLutHi;
             Vector256<sbyte> lutLo            = s_avx_decodeLutLo;
             Vector256<sbyte> lutShift         = s_avx_decodeLutShift;
-            Vector256<sbyte> mask5F           = s_avx_decodeMask5F;
-            Vector256<sbyte> shift5F          = Avx.SetAllVector256<sbyte>(33);     // high nibble is 0x5 -> range 'P' .. 'Z' for shift, not '+' (0x2)
+            Vector256<sbyte> mask2F           = s_avx_decodeMask2F;
             Vector256<sbyte> shuffleConstant0 = Avx.StaticCast<int, sbyte>(Avx.SetAllVector256(0x01400140));
             Vector256<short> shuffleConstant1 = Avx.StaticCast<int, short>(Avx.SetAllVector256(0x00011000));
             Vector256<sbyte> shuffleVec       = s_avx_decodeShuffleVec;
@@ -274,24 +334,20 @@ namespace gfoidl.Base64
                     throw new NotSupportedException(); // just in case new types are introduced in the future
                 }
 
-                Vector256<sbyte> hiNibbles = Avx2.And(Avx.StaticCast<int, sbyte>(Avx2.ShiftRightLogical(Avx.StaticCast<sbyte, int>(str), 4)), mask5F);
-                Vector256<sbyte> lowerBound = Avx2.Shuffle(lutLo, hiNibbles);
-                Vector256<sbyte> upperBound = Avx2.Shuffle(lutHi, hiNibbles);
+                Vector256<sbyte> hiNibbles = Avx2.And(Avx.StaticCast<int, sbyte>(Avx2.ShiftRightLogical(Avx.StaticCast<sbyte, int>(str), 4)), mask2F);
+                Vector256<sbyte> loNibbles = Avx2.And(str, mask2F);
+                Vector256<sbyte> hi        = Avx2.Shuffle(lutHi, hiNibbles);
+                Vector256<sbyte> lo        = Avx2.Shuffle(lutLo, loNibbles);
 
-                Vector256<sbyte> below   = Avx2Helper.LessThan(str, lowerBound, allOnes);
-                Vector256<sbyte> above   = Avx2.CompareGreaterThan(str, upperBound);
-                Vector256<sbyte> eq5F    = Avx2.CompareEqual(str, mask5F);
-                Vector256<sbyte> outside = Avx2.AndNot(eq5F, Avx2.Or(below, above));
-
-                if (Avx2.MoveMask(outside) != 0)
+                if (!Avx.TestZ(lo, hi))
                 {
                     success = false;
                     break;
                 }
 
-                Vector256<sbyte> shift = Avx2.Shuffle(lutShift, hiNibbles);
+                Vector256<sbyte> eq2F  = Avx2.CompareEqual(str, mask2F);
+                Vector256<sbyte> shift = Avx2.Shuffle(lutShift, Avx2.Add(eq2F, hiNibbles));
                 str                    = Avx2.Add(str, shift);
-                str                    = Avx2.Add(str, Avx2.And(eq5F, shift5F));
 
                 Vector256<short> merge_ab_and_bc = Avx2.MultiplyAddAdjacent(Avx.StaticCast<sbyte, byte>(str), shuffleConstant0);
                 Vector256<int> @out              = Avx2.MultiplyAddAdjacent(merge_ab_and_bc, shuffleConstant1);
@@ -337,8 +393,7 @@ namespace gfoidl.Base64
             Vector128<sbyte> lutHi            = s_sse_decodeLutHi;
             Vector128<sbyte> lutLo            = s_sse_decodeLutLo;
             Vector128<sbyte> lutShift         = s_sse_decodeLutShift;
-            Vector128<sbyte> mask5F           = s_sse_decodeMask5F;
-            Vector128<sbyte> shift5F          = Sse2.SetAllVector128((sbyte)33); // high nibble is 0x5 -> range 'P' .. 'Z' for shift, not '+' (0x2)
+            Vector128<sbyte> mask2F           = s_sse_decodeMask2F;
             Vector128<sbyte> shuffleConstant0 = Sse.StaticCast<int, sbyte>(Sse2.SetAllVector128(0x01400140));
             Vector128<short> shuffleConstant1 = Sse.StaticCast<int, short>(Sse2.SetAllVector128(0x00011000));
             Vector128<sbyte> shuffleVec       = s_sse_decodeShuffleVec;
@@ -364,29 +419,31 @@ namespace gfoidl.Base64
                     throw new NotSupportedException(); // just in case new types are introduced in the future
                 }
 
-                Vector128<sbyte> hiNibbles = Sse2.And(Sse.StaticCast<int, sbyte>(Sse2.ShiftRightLogical(Sse.StaticCast<sbyte, int>(str), 4)), mask5F);
-                Vector128<sbyte> lowerBound = Ssse3.Shuffle(lutLo, hiNibbles);
-                Vector128<sbyte> upperBound = Ssse3.Shuffle(lutHi, hiNibbles);
-
-                Vector128<sbyte> below   = Sse2.CompareLessThan(str, lowerBound);
-                Vector128<sbyte> above   = Sse2.CompareGreaterThan(str, upperBound);
-                Vector128<sbyte> eq5F    = Sse2.CompareEqual(str, mask5F);
-                Vector128<sbyte> outside = Sse2.AndNot(eq5F, Sse2.Or(below, above));
-
-                if (Sse2.MoveMask(outside) != 0)
+                Vector128<sbyte> hiNibbles = Sse2.And(Sse.StaticCast<int, sbyte>(Sse2.ShiftRightLogical(Sse.StaticCast<sbyte, int>(str), 4)), mask2F);
+                Vector128<sbyte> loNibbles = Sse2.And(str, mask2F);
+                Vector128<sbyte> hi        = Ssse3.Shuffle(lutHi, hiNibbles);
+                Vector128<sbyte> lo        = Ssse3.Shuffle(lutLo, loNibbles);
+#if NETCOREAPP3_0
+                // https://github.com/dotnet/coreclr/issues/21130
+                //Vector128<sbyte> zero = Vector128<sbyte>.Zero;
+                Vector128<sbyte> zero = Sse2.SetZeroVector128<sbyte>();
+#elif NETCOREAPP2_1
+                Vector128<sbyte> zero = Sse2.SetZeroVector128<sbyte>();
+#endif
+                if (Sse2.MoveMask(Sse2.CompareGreaterThan(Sse2.And(lo, hi), zero)) != 0)
                 {
                     success = false;
                     break;
                 }
 
-                Vector128<sbyte> shift = Ssse3.Shuffle(lutShift, hiNibbles);
+                Vector128<sbyte> eq2F  = Sse2.CompareEqual(str, mask2F);
+                Vector128<sbyte> shift = Ssse3.Shuffle(lutShift, Sse2.Add(eq2F, hiNibbles));
                 str                    = Sse2.Add(str, shift);
-                str                    = Sse2.Add(str, Sse2.And(eq5F, shift5F));
 
                 Vector128<short> merge_ab_and_bc = Ssse3.MultiplyAddAdjacent(Sse.StaticCast<sbyte, byte>(str), shuffleConstant0);
 #if NETCOREAPP3_0
                 Vector128<int> @out = Sse2.MultiplyAddAdjacent(merge_ab_and_bc, shuffleConstant1);
-#else
+#elif NETCOREAPP2_1
                 Vector128<int> @out = Sse2.MultiplyHorizontalAdd(merge_ab_and_bc, shuffleConstant1);
 #endif
                 str = Ssse3.Shuffle(Sse.StaticCast<int, sbyte>(@out), shuffleVec);
@@ -395,7 +452,7 @@ namespace gfoidl.Base64
                 // https://github.com/dotnet/coreclr/issues/21132
                 Unsafe.As<byte, Vector128<sbyte>>(ref destBytes) = str;
 
-                src       = ref Unsafe.Add(ref src, 16);
+                src       = ref Unsafe.Add(ref src,  16);
                 destBytes = ref Unsafe.Add(ref destBytes, 12);
             }
 
@@ -413,154 +470,37 @@ namespace gfoidl.Base64
         }
 #endif
         //---------------------------------------------------------------------
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static int DecodeThree<T>(ref T encoded, ref sbyte decodingMap)
-        {
-            uint t0, t1, t2;
-
-            if (typeof(T) == typeof(byte))
-            {
-                ref byte tmp = ref Unsafe.As<T, byte>(ref encoded);
-                t0 = Unsafe.Add(ref tmp, 0);
-                t1 = Unsafe.Add(ref tmp, 1);
-                t2 = Unsafe.Add(ref tmp, 2);
-            }
-            else if (typeof(T) == typeof(char))
-            {
-                ref char tmp = ref Unsafe.As<T, char>(ref encoded);
-                t0 = Unsafe.Add(ref tmp, 0);
-                t1 = Unsafe.Add(ref tmp, 1);
-                t2 = Unsafe.Add(ref tmp, 2);
-            }
-            else
-            {
-                throw new NotSupportedException();  // just in case new types are introduced in the future
-            }
-
-            int i0 = Unsafe.Add(ref decodingMap, (IntPtr)t0);
-            int i1 = Unsafe.Add(ref decodingMap, (IntPtr)t1);
-            int i2 = Unsafe.Add(ref decodingMap, (IntPtr)t2);
-
-            return i0 << 18 | i1 << 12 | i2 << 6;
-        }
-        //---------------------------------------------------------------------
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static int DecodeTwo<T>(ref T encoded, ref sbyte decodingMap)
-        {
-            uint t0, t1;
-
-            if (typeof(T) == typeof(byte))
-            {
-                ref byte tmp = ref Unsafe.As<T, byte>(ref encoded);
-                t0           = Unsafe.Add(ref tmp, 0);
-                t1           = Unsafe.Add(ref tmp, 1);
-            }
-            else if (typeof(T) == typeof(char))
-            {
-                ref char tmp = ref Unsafe.As<T, char>(ref encoded);
-                t0 = Unsafe.Add(ref tmp, 0);
-                t1 = Unsafe.Add(ref tmp, 1);
-            }
-            else
-            {
-                throw new NotSupportedException();  // just in case new types are introduced in the future
-            }
-
-            int i0 = Unsafe.Add(ref decodingMap, (IntPtr)t0);
-            int i1 = Unsafe.Add(ref decodingMap, (IntPtr)t1);
-
-            return i0 << 18 | i1 << 12;
-        }
-        //---------------------------------------------------------------------
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static void WriteTwoLowOrderBytes(ref byte destination, uint destIndex, int value)
-        {
-            Unsafe.Add(ref destination, (IntPtr)(destIndex + 0)) = (byte)(value >> 16);
-            Unsafe.Add(ref destination, (IntPtr)(destIndex + 1)) = (byte)(value >> 8);
-        }
-        //---------------------------------------------------------------------
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static void WriteOneLowOrderByte(ref byte destination, uint destIndex, int value)
-        {
-            Unsafe.Add(ref destination, (IntPtr)destIndex) = (byte)(value >> 16);
-        }
-        //---------------------------------------------------------------------
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static int GetDataLen(int urlEncodedLen, out int base64Len, bool isFinalBlock = true)
-        {
-            if (isFinalBlock)
-            {
-                // Shortcut for Guid and other 16 byte data
-                if (urlEncodedLen == 22)
-                {
-                    base64Len = 24;
-                    return 16;
-                }
-
-                int numPaddingChars = GetNumBase64PaddingCharsToAddForDecode(urlEncodedLen);
-                base64Len           = urlEncodedLen + numPaddingChars;
-
-                Debug.Assert(base64Len % 4 == 0, "Invariant: Array length must be a multiple of 4.");
-
-                int dataLength = (base64Len >> 2) * 3 - numPaddingChars;
-                return dataLength;
-            }
-            else
-            {
-                base64Len = urlEncodedLen;
-                return (urlEncodedLen >> 2) * 3;
-            }
-        }
-        //---------------------------------------------------------------------
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static int GetNumBase64PaddingCharsToAddForDecode(int urlEncodedLen)
-        {
-            // Calculation is:
-            // switch (inputLength % 4)
-            // 0 -> 0
-            // 2 -> 2
-            // 3 -> 1
-            // default -> format exception
-
-            int result = (4 - urlEncodedLen) & 3;
-
-            if (result == 3)
-                ThrowHelper.ThrowMalformedInputException(urlEncodedLen);
-
-            return result;
-        }
-        //---------------------------------------------------------------------
 #if NETCOREAPP
         private static readonly Vector128<sbyte> s_sse_decodeLutLo;
         private static readonly Vector128<sbyte> s_sse_decodeLutHi;
         private static readonly Vector128<sbyte> s_sse_decodeLutShift;
-        private static readonly Vector128<sbyte> s_sse_decodeMask5F;
+        private static readonly Vector128<sbyte> s_sse_decodeMask2F;
 
 #if NETCOREAPP3_0
         private static readonly Vector256<sbyte> s_avx_decodeLutLo;
         private static readonly Vector256<sbyte> s_avx_decodeLutHi;
         private static readonly Vector256<sbyte> s_avx_decodeLutShift;
-        private static readonly Vector256<sbyte> s_avx_decodeMask5F;
+        private static readonly Vector256<sbyte> s_avx_decodeMask2F;
 #endif
 #endif
         // internal because tests use this map too
         internal static readonly sbyte[] s_decodingMap = {
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 62, -1, -1,
-            52, 53, 54, 55, 56, 57, 58, 59, 60, 61, -1, -1, -1, -1, -1, -1,
-            -1,  0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14,
-            15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, -1, -1, -1, -1, 63,
+            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 62, -1, -1, -1, 63,         //62 is placed at index 43 (for +), 63 at index 47 (for /)
+            52, 53, 54, 55, 56, 57, 58, 59, 60, 61, -1, -1, -1, -1, -1, -1,         //52-61 are placed at index 48-57 (for 0-9), 64 at index 61 (for =)
+            -1, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14,
+            15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, -1, -1, -1, -1, -1,         //0-25 are placed at index 65-90 (for A-Z)
             -1, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40,
-            41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, -1, -1, -1, -1, -1,
+            41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, -1, -1, -1, -1, -1,         //26-51 are placed at index 97-122 (for a-z)
+            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,         // Bytes over 122 ('z') are invalid and cannot be decoded
+            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,         // Hence, padding the map with 255, which indicates invalid input
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1
         };
     }
 }
