+++ source/gfoidl.Base64/Internal/Base64Encoder/Base64Encoder.cs
@@ -2,7 +2,7 @@
 using System.Runtime.Intrinsics.X86;
 #endif
 
-namespace gfoidl.Base64
+namespace gfoidl.Base64.Internal
 {
     public sealed partial class Base64Encoder : Base64EncoderImpl
     {
