+++ source/gfoidl.Base64/Internal/Base64UrlEncoder/Base64UrlEncoder.cs
@@ -3,7 +3,7 @@ using System.Runtime.Intrinsics;
 using System.Runtime.Intrinsics.X86;
 #endif
 
-namespace gfoidl.Base64
+namespace gfoidl.Base64.Internal
 {
     public sealed partial class Base64UrlEncoder : Base64EncoderImpl
     {
