+++ source/gfoidl.Base64/Internal/Base64UrlEncoder/Lookup-Table.xlsx
