+++ source/gfoidl.Base64/gfoidl.Base64.csproj
@@ -36,7 +36,7 @@ For .NET Core 2.1+ encoding / decoding is done with SIMD-support.
     </ItemGroup>
 
     <ItemGroup Condition="'$(TargetFramework)' != 'netcoreapp3.0'">
-        <Compile Remove="Avx2Helper.cs" />
+        <Compile Remove="Internal\Avx2Helper.cs" />
     </ItemGroup>
 
     <ItemGroup>
