+++ tests/gfoidl.Base64.Tests/Internal/Avx2HelperTests/Write.cs
@@ -1,8 +1,9 @@
 ﻿using System.Runtime.CompilerServices;
 using System.Runtime.Intrinsics;
+using gfoidl.Base64.Internal;
 using NUnit.Framework;
 
-namespace gfoidl.Base64.Tests.Avx2HelperTests
+namespace gfoidl.Base64.Tests.Internal.Avx2HelperTests
 {
     [TestFixture]
     public class Write
