+++ tests/gfoidl.Base64.Tests/Internal/Base64EncoderTests/Roundtrips.cs
@@ -2,9 +2,10 @@
 using System.Buffers;
 using System.Runtime.InteropServices;
 using System.Text;
+using gfoidl.Base64.Internal;
 using NUnit.Framework;
 
-namespace gfoidl.Base64.Tests.Base64EncoderTests
+namespace gfoidl.Base64.Tests.Internal.Base64EncoderTests
 {
     [TestFixture(typeof(byte))]
     [TestFixture(typeof(char))]
