+++ tests/gfoidl.Base64.Tests/Internal/Base64UrlEncoderTests/EncodingMaps.cs
@@ -1,12 +1,13 @@
 ﻿using System;
+using gfoidl.Base64.Internal;
 using NUnit.Framework;
 
-namespace gfoidl.Base64.Tests.Base64EncoderTests
+namespace gfoidl.Base64.Tests.Internal.Base64UrlEncoderTests
 {
     [TestFixture]
     public class EncodingMaps
     {
-        private static readonly string s_characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
+        private static readonly string s_characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_";
         //---------------------------------------------------------------------
         [Test]
         public void Verify_encoding_map()
@@ -16,7 +17,7 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
             for (int i = 0; i < s_characters.Length; ++i)
                 data[i] = (byte)s_characters[i];
 
-            CollectionAssert.AreEqual(Base64Encoder.s_encodingMap, data);
+            CollectionAssert.AreEqual(Base64UrlEncoder.s_encodingMap, data);
         }
         //---------------------------------------------------------------------
         [Test]
@@ -28,7 +29,7 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
             for (int i = 0; i < s_characters.Length; ++i)
                 data[s_characters[i]] = (sbyte)i;
 
-            CollectionAssert.AreEqual(Base64Encoder.s_decodingMap, data);
+            CollectionAssert.AreEqual(Base64UrlEncoder.s_decodingMap, data);
         }
     }
 }
