+++ tests/gfoidl.Base64.Tests/Internal/Base64UrlEncoderTests/GetDecodedLength.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Text;
+using gfoidl.Base64.Internal;
 using NUnit.Framework;
 
-namespace gfoidl.Base64.Tests.Base64EncoderTests
+namespace gfoidl.Base64.Tests.Internal.Base64UrlEncoderTests
 {
     [TestFixture]
     public class GetDecodedLength
@@ -10,7 +11,7 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
         [Test]
         public void EncodedLength_is_0___0()
         {
-            var sut = new Base64Encoder();
+            var sut = new Base64UrlEncoder();
 
             int actual = sut.GetDecodedLength(0);
 
@@ -20,14 +21,14 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
         [Test]
         public void EncodedSpan_1_to_50_byte_given___correct_decoded_len()
         {
-            var sut = new Base64Encoder();
+            var sut = new Base64UrlEncoder();
 
             Assert.Multiple(() =>
             {
                 for (int i = 1; i < 50; ++i)
                 {
                     var data                     = new byte[i];
-                    string base64WoPaddingString = Convert.ToBase64String(data);
+                    string base64WoPaddingString = Convert.ToBase64String(data).TrimEnd('=');
                     byte[] base64WoPadding       = Encoding.ASCII.GetBytes(base64WoPaddingString);
 
                     int actual = sut.GetDecodedLength(base64WoPadding);
@@ -40,14 +41,14 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
         [Test]
         public void EncodedSpan_1_to_50_char_given___correct_decoded_len()
         {
-            var sut = new Base64Encoder();
+            var sut = new Base64UrlEncoder();
 
             Assert.Multiple(() =>
             {
                 for (int i = 1; i < 50; ++i)
                 {
                     var data                     = new byte[i];
-                    string base64WoPaddingString = Convert.ToBase64String(data);
+                    string base64WoPaddingString = Convert.ToBase64String(data).TrimEnd('=');
 
                     int actual = sut.GetDecodedLength(base64WoPaddingString.AsSpan());
 
@@ -57,21 +58,29 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
         }
         //---------------------------------------------------------------------
         [Test]
-        public void EncodedLength_is_int_Max___OK()
+        public void EncodedLength_is_int_Max___throws_ArgumentOutOfRange()
         {
-            var sut = new Base64Encoder();
+            var sut = new Base64UrlEncoder();
 
-            int actual = sut.GetDecodedLength(int.MaxValue);
-
-            Assert.AreEqual(int.MaxValue / 4 * 3, actual);
+            Assert.Throws<ArgumentOutOfRangeException>(() => sut.GetDecodedLength(int.MaxValue));
         }
         //---------------------------------------------------------------------
         [Test]
         public void EncodedLength_is_negative___throws_ArgumentOutOfRange()
         {
-            var sut = new Base64Encoder();
+            var sut = new Base64UrlEncoder();
 
             Assert.Throws<ArgumentOutOfRangeException>(() => sut.GetDecodedLength(-1));
+        }
+        //---------------------------------------------------------------------
+        [Test]
+        [TestCase(1)]
+        [TestCase(5)]
+        public void Malformed_input_length___throws_FormatException(int len)
+        {
+            var sut = new Base64UrlEncoder();
+
+            Assert.Throws<FormatException>(() => sut.GetDecodedLength(len));
         }
     }
 }
