+++ tests/gfoidl.Base64.Tests/Internal/Base64UrlEncoderTests/GetEncodedLength.cs
@@ -1,7 +1,8 @@
 ﻿using System;
+using gfoidl.Base64.Internal;
 using NUnit.Framework;
 
-namespace gfoidl.Base64.Tests.Base64EncoderTests
+namespace gfoidl.Base64.Tests.Internal.Base64UrlEncoderTests
 {
     [TestFixture]
     public class GetEncodedLength
@@ -9,7 +10,7 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
         [Test]
         public void SourceLength_is_0___0()
         {
-            var sut = new Base64Encoder();
+            var sut = new Base64UrlEncoder();
 
             int actual = sut.GetEncodedLength(0);
 
@@ -19,14 +20,14 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
         [Test]
         public void SourceLength_1_to_50___correct_encoded_len()
         {
-            var sut = new Base64Encoder();
+            var sut = new Base64UrlEncoder();
 
             Assert.Multiple(() =>
             {
                 for (int i = 1; i < 50; ++i)
                 {
-                    var data               = new byte[i];
-                    string base64WoPadding = Convert.ToBase64String(data);
+                    var data = new byte[i];
+                    string base64WoPadding = Convert.ToBase64String(data).TrimEnd('=');
 
                     int actual = sut.GetEncodedLength(i);
 
@@ -38,7 +39,7 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
         [Test]
         public void SourceLength_gt_MaximumEncodeLength___throws_ArgumentOutOfRange()
         {
-            var sut = new Base64Encoder();
+            var sut = new Base64UrlEncoder();
 
             Assert.Throws<ArgumentOutOfRangeException>(() => sut.GetEncodedLength(Base64Encoder.MaximumEncodeLength + 1));
         }
@@ -46,7 +47,7 @@ namespace gfoidl.Base64.Tests.Base64EncoderTests
         [Test]
         public void SourceLength_is_negative___throws_ArgumentOutOfRange()
         {
-            var sut = new Base64Encoder();
+            var sut = new Base64UrlEncoder();
 
             Assert.Throws<ArgumentOutOfRangeException>(() => sut.GetEncodedLength(-1));
         }
